--- community/command-line/src/main/resources/META-INF/services/org.neo4j.commandline.admin.AdminCommand$Provider
@@ -1,1 +1,0 @@
-org.neo4j.commandline.admin.ImportCommand$Provider
