+++ community/dbms/src/main/resources/META-INF/services/org.neo4j.commandline.admin.AdminCommand$Provider
@@ -1,0 +1,1 @@
+org.neo4j.commandline.dbms.ImportCommand$Provider
